detach loop from backend
