+++ examples/cube_glfw.py
@@ -5,7 +5,7 @@ Cube glfw
 Run a wgpu example on the glfw backend.
 """
 
-from rendercanvas.glfw import RenderCanvas, run
+from rendercanvas.glfw import RenderCanvas, loop
 
 from rendercanvas.utils.cube import setup_drawing_sync
 
@@ -18,4 +18,4 @@ canvas.request_draw(draw_frame)
 
 
 if __name__ == "__main__":
-    run()
+    loop.run()
