+++ rendercanvas/_scheduler.py
@@ -1,0 +1,205 @@
+"""
+The scheduler class/loop.
+"""
+
+import time
+import weakref
+
+from ._coreutils import BaseEnum
+from ._async_sniffs import sleep, Event  # todo: rename this module
+
+
+class UpdateMode(BaseEnum):
+    """The different modes to schedule draws for the canvas."""
+
+    manual = None  #: Draw events are never scheduled. Draws only happen when you ``canvas.force_draw()``, and maybe when the GUI system issues them (e.g. when resizing).
+    ondemand = None  #: Draws are only scheduled when ``canvas.request_draw()`` is called when an update is needed. Safes your laptop battery. Honours ``min_fps`` and ``max_fps``.
+    continuous = None  #: Continuously schedules draw events, honouring ``max_fps``. Calls to ``canvas.request_draw()`` have no effect.
+    fastest = None  #: Continuously schedules draw events as fast as possible. Gives high FPS (and drains your battery).
+
+
+class Scheduler:
+    """Helper class to schedule event processing and drawing."""
+
+    # This class makes the canvas tick. Since we do not own the event-loop, but
+    # ride on e.g. Qt, asyncio, wx, JS, or something else, our little "loop" is
+    # implemented with a timer.
+    #
+    # The loop looks a little like this:
+    #
+    #     ________________      __      ________________      __      rd = request_draw
+    #   /   wait           \  / rd \  /   wait           \  / rd \
+    #  |                    ||      ||                    ||      |
+    # --------------------------------------------------------------------> time
+    #  |                    |       |                     |       |
+    #  schedule             tick    draw                  tick    draw
+    #
+    # With update modes 'ondemand' and 'manual', the loop ticks at the same rate
+    # as on 'continuous' mode, but won't draw every tick:
+    #
+    #     ________________     ________________      __
+    #   /    wait          \  /   wait          \  / rd \
+    #  |                    ||                   ||      |
+    # --------------------------------------------------------------------> time
+    #  |                    |                    |       |
+    #  schedule             tick                tick     draw
+    #
+    # A tick is scheduled by calling _schedule_next_tick(). If this method is
+    # called when the timer is already running, it has no effect. In the _tick()
+    # method, events are processed (including animations). Then, depending on
+    # the mode and whether a draw was requested, a new tick is scheduled, or a
+    # draw is requested. In the latter case, the timer is not started, but we
+    # wait for the canvas to perform a draw. In _draw_drame_and_present() the
+    # draw is done, and a new tick is scheduled.
+    #
+    # The next tick is scheduled when a draw is done, and not earlier, otherwise
+    # the drawing may not keep up with the ticking.
+    #
+    # On desktop canvases the draw usually occurs very soon after it is
+    # requested, but on remote frame buffers, it may take a bit longer. To make
+    # sure the rendered image reflects the latest state, these backends may
+    # issue an extra call to _process_events() right before doing the draw.
+    #
+    # When the window is minimized, the draw will not occur until the window is
+    # shown again. For the canvas to detect minimized-state, it will need to
+    # receive GUI events. This is one of the reasons why the loop object also
+    # runs a timer-loop.
+    #
+    # The drawing itself may take longer than the intended wait time. In that
+    # case, it will simply take longer than we hoped and get a lower fps.
+    #
+    # Note that any extra draws, e.g. via force_draw() or due to window resizes,
+    # don't affect the scheduling loop; they are just extra draws.
+
+    def __init__(self, canvas, events, loop, *, mode="ondemand", min_fps=1, max_fps=30):
+        self.name = f"{canvas.__class__.__name__} scheduler"
+        assert loop is not None
+
+        # We don't keep a ref to the canvas to help gc. This scheduler object can be
+        # referenced via a callback, but it won't prevent the canvas from being deleted!
+        self._canvas_ref = weakref.ref(canvas)
+        self._events = events
+        # ... = canvas.get_context() -> No, context creation should be lazy!
+
+        # Scheduling variables
+        if mode not in UpdateMode:
+            raise ValueError(
+                f"Invalid update_mode '{mode}', must be in {set(UpdateMode)}."
+            )
+        self._mode = mode
+        self._min_fps = float(min_fps)
+        self._max_fps = float(max_fps)
+        self._draw_requested = True  # Start with a draw in ondemand mode
+        self._last_draw_time = 0
+        self._async_draw_event = None
+
+        # Keep track of fps
+        self._draw_stats = 0, time.perf_counter()
+
+        self._last_tick_time = -0.1
+        # todo: some vars maybe moved to inside the coro
+
+        loop.add_scheduler(self)
+        loop.add_task(self.__scheduler_task)
+
+    def get_canvas(self):
+        """Get the canvas, or None if it is closed or gone."""
+        canvas = self._canvas_ref()
+        if canvas is None or canvas.get_closed():
+            return None
+        else:
+            return canvas
+
+    def request_draw(self):
+        """Request a new draw to be done. Only affects the 'ondemand' mode."""
+        # Just set the flag
+        self._draw_requested = True
+
+    async def __scheduler_task(self):
+        """The coro that reprsents the scheduling loop for a canvas."""
+
+        while True:
+            # Determine delay
+            if self._mode == "fastest" or self._max_fps <= 0:
+                delay = 0
+            else:
+                delay = 1 / self._max_fps
+                delay = 0 if delay < 0 else delay  # 0 means cannot keep up
+
+            # Offset delay for time spent on processing events, etc.
+            time_since_tick_start = time.perf_counter() - self._last_tick_time
+            delay -= time_since_tick_start
+            delay = max(0, delay)
+
+            # Wait
+            await sleep(delay)
+
+            # Get canvas or stop
+            if (canvas := self.get_canvas()) is None:
+                break
+
+            # Below is the "tick"
+
+            self._last_tick_time = time.perf_counter()
+
+            # Process events, handlers may request a draw
+            await canvas._process_events()
+
+            # Determine what to do next ...
+
+            do_draw = False
+
+            if self._mode == "fastest":
+                # fastest: draw continuously as fast as possible, ignoring fps settings.
+                do_draw = True
+            elif self._mode == "continuous":
+                # continuous: draw continuously, aiming for a steady max framerate.
+                do_draw = True
+            elif self._mode == "ondemand":
+                # ondemand: draw when needed (detected by calls to request_draw).
+                # Aim for max_fps when drawing is needed, otherwise min_fps.
+                if self._draw_requested:
+                    do_draw = True
+                elif (
+                    self._min_fps > 0
+                    and time.perf_counter() - self._last_draw_time > 1 / self._min_fps
+                ):
+                    do_draw = True
+
+            elif self._mode == "manual":
+                pass
+            else:
+                raise RuntimeError(f"Unexpected scheduling mode: '{self._mode}'")
+
+            # If we don't want to draw, we move to the next iter
+            if not do_draw:
+                continue
+
+            canvas._rc_request_draw()
+            self._async_draw_event = Event()
+            await self._async_draw_event.wait()
+
+        await self._events._rc_canvas_close()
+
+    def on_draw(self):
+        # Bookkeeping
+        self._last_draw_time = time.perf_counter()
+        self._draw_requested = False
+
+        # Keep ticking
+        if self._async_draw_event:
+            self._async_draw_event.set()
+            self._async_draw_event = None
+
+        # Update stats
+        count, last_time = self._draw_stats
+        count += 1
+        if time.perf_counter() - last_time > 1.0:
+            fps = count / (time.perf_counter() - last_time)
+            self._draw_stats = 0, time.perf_counter()
+        else:
+            fps = None
+            self._draw_stats = count, last_time
+
+        # Return fps or None. Will change with better stats at some point
+        return fps
